S15.f.8: extract plugin.ts Foundry + http_fetch tool registrations (5071 → 4323 LOC)
Eighth sub-slice of the multi-PR train decomposing cli/src/plugin.ts
toward the Phase 2 §4.2 cap of 3000 LOC. Lifts the 10-tool Foundry +
http_fetch cluster from inline initAGT() registrations into a new
cli/src/core/agt-tools/ directory.

LOC delta:
- cli/src/plugin.ts: 5071 → 4323 (−748)
- cli/src/core/safe-json.ts: NEW (12 LOC)
- cli/src/core/agt-tools/http-fetch.ts: NEW (39 LOC)
- cli/src/core/agt-tools/foundry.ts: NEW (763 LOC)
- Cumulative S15.f: 7139 → 4323 (−2816, 78% to cap)

Tools moved (bodies byte-identical):
- http_fetch → core/agt-tools/http-fetch.ts
- foundry_code_execute, foundry_image_generation, foundry_web_search,
  foundry_file_search, foundry_memory, foundry_conversations,
  foundry_evaluations, foundry_deployments, foundry_agents
  → core/agt-tools/foundry.ts (registerFoundryTools)

FoundryToolsDeps threads (log, config, getFoundryProject) — the
late-bound accessor preserves the previous semantics where tool
execution reads the current foundryProject value (initFoundry runs
fire-and-forget alongside initAGT in register()).

safeJson(obj, maxLen) lifted from inline register()-scope helper to
a module-level utility in core/safe-json.ts.

No tool surface changes: names, descriptions, parameter schemas,
execute return shapes byte-identical. Egress posture unchanged
(http_fetch still routes via 127.0.0.1:8443 router proxy). Foundry
router endpoints (paths, bodies, timeouts) all unchanged.

Verification:
- npx tsc --noEmit clean
- npm run lint: 30 warnings (unchanged)
- npm test: 454 pass / 2 skipped (unchanged)
- npm run build: clean

Audit: docs/security-audits/2026-04-29-phase2-hotspot-plugin-cli-f8.md
